Moves from value endpoints to events.
Values are simpler and make a certain logical sense. Events require we think about actions instead of streams, and often require we manually cache state.

But events are no-ops when not in use whereas values seem to require something close to per-frame computation. Switching to events more than halved the CPU load of the sample envelopes, so it’s an easy decision.